Instance-aware API
Rather than falling back to the stored API URL, this now explicitly relies on the stored session data, forcing it to be instance aware, addressing [this](https://github.com/smtech/canvas-management/issues/12) and [this](https://github.com/smtech/canvashack/issues/8), although there is an upcoming commit to [canvas-management](https://github.com/smtech/canvas-management) that will need to be promulgated to better demonstrate how to get the API URL from the LTI request.

The core issue is that using LTI_Tool_Provider to authenticate and route LTI requests is a badly architected mess, and that dependency should be ripped out and replaced with something simpler that yields exactly the information needed, reliably.